+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [13.6.1](https://github.com/gravity-ui/markdown-editor/compare/v13.6.0...v13.6.1) (2024-08-01)
+
+
+### Bug Fixes
+
+* **build:** fixed build of styles ([#317](https://github.com/gravity-ui/markdown-editor/issues/317)) ([0349c71](https://github.com/gravity-ui/markdown-editor/commit/0349c71b597cab9ce72679d1854bb7e0357daaec))
+
+## [13.6.0](https://github.com/gravity-ui/markdown-editor/compare/v13.5.3...v13.6.0) (2024-08-01)
+
+
+### Features
+
+* add FoldingHeading extension ([#314](https://github.com/gravity-ui/markdown-editor/issues/314)) ([b904704](https://github.com/gravity-ui/markdown-editor/commit/b9047045c77656e4721eb1852f8d2ec018d9f5db))
+
+
+### Bug Fixes
+
+* **bundle:** scroll to current line after moving cursor in markup mode ([#313](https://github.com/gravity-ui/markdown-editor/issues/313)) ([dd56733](https://github.com/gravity-ui/markdown-editor/commit/dd567333865914d7be1de734cb4413737e2a9b62))
+
 ## [13.5.3](https://github.com/gravity-ui/markdown-editor/compare/v13.5.2...v13.5.3) (2024-07-29)